Added code to update manager, to remove backup (.bak) from update, after 3 days.

+++ src/WOWCAM/WOWCAM.Core/DefaultUpdateManager.cs
@@ -144,7 +144,11 @@ namespace WOWCAM.Core
 
                 if (File.Exists(bakFilePath))
                 {
-                    File.Delete(bakFilePath);
+                    var fileInfo = new FileInfo(bakFilePath);
+                    if (fileInfo.LastAccessTimeUtc < DateTime.UtcNow.AddDays(-3))
+                    {
+                        File.Delete(bakFilePath);
+                    }
                 }
 
                 return !File.Exists(bakFilePath);
